Refactor once more, moved stuff from main.py to application.py
